Internalized package-resolver dependency
