+++ package.json
@@ -55,7 +55,6 @@
     "glob": "^8.0.3",
     "locate-path": "^7.1.1",
     "on-change": "^4.0.1",
-    "package-resolver": "^1.1.0",
     "progress": "^2.0.3",
     "tmp": "^0.2.1",
     "yaml-cfn": "^0.3.2",
